fix memory leaks

+++ src/Tiles.h
@@ -67,5 +67,5 @@ public:
 	//// Add whole new row of tile. Only for main menu.
 	static void AddTileRow(std::vector < Tiles>& tile, TileType type, const int count, const f32 width, const f32 height, const AEVec2 pos);
 
-	static void CreateDialogue(int count, AEVec2 tilePos);
+	static void CreateDialogue(short count, AEVec2 tilePos);
 };
